The project's directory structure has been reorganized. see changes.md for details.
