+++ include/shoot/bullet_solver.hpp
