+++ log/fric_log.txt
