+++ src/device/dji_motor.cc
